// shogihome/src/server/book/proto/sbk.ts
export interface SBook {
  Author: string;
  Description: string;
  BookStates: SBookState[];
}

/** 局面 */
export interface SBookState {
  /** ID */
  Id: number;
  /** 局面ハッシュ */
  BoardKey: bigint;
  /** 持ち駒　手番 */
  HandKey: number;
  /** 対局数 */
  Games: number;
  /** 先手勝ち数 */
  WonBlack: number;
  /** 後手勝ち数 */
  WonWhite: number;
  /** sfen形式のposition情報 */
  Position?: string | undefined;
  /** コメント */
  Comment?: string | undefined;
  /** この局面における定跡 */
  Moves: SBookMove[];
  /** 評価値 */
  Evals: SBookEval[];
}

/** 指し手 */
export interface SBookMove {
  /**
   * 指し手
   * move
   * 0-3 from dan  (1-9)  打つ手の場合は0
   * 4-7 form suji (1-9)  打つ手の場合は0
   * 8-11 to dan (1-9)
   * 12-15 to suji (1-9)
   * 19    promotion flag
   * 20-23 capture piece 取った駒 (Promotionフラグも含む
   * 24-29 piece
   * 31   0:black 1:white
   */
  Move: number;
  /** さして評価 */
  Evaluation: SBookMoveEvaluation;
  /** 重み */
  Weight: number;
  /** 次の局面 */
  NextStateId: number;
}

/** 評価値 */
export interface SBookEval {
  /** 評価値 */
  EvaluationValue: number;
  /** 深さ */
  Depth: number;
  SelDepth: number;
  /** ノード数 */
  Nodes: bigint;
  /** 読み筋 */
  Variation: string;
  /** エンジン名 */
  EngineName?: string | undefined;
}

function createBaseSBook(): SBook {
  return { Author: "", Description: "", BookStates: [] };
}

export const SBook: MessageFns<SBook> = {
  encode(message: SBook, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.Author !== "") {
      writer.uint32(10).string(message.Author);
    }
    if (message.Description !== "") {
      writer.uint32(18).string(message.Description);
    }
    for (const v of message.BookStates) {
      SBookState.encode(v!, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SBook {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSBook();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.Author = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.Description = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.BookStates.push(SBookState.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SBook {
    return {
      Author: isSet(object.Author) ? globalThis.String(object.Author) : "",
      Description: isSet(object.Description) ? globalThis.String(object.Description) : "",
      BookStates: globalThis.Array.isArray(object?.BookStates)
        ? object.BookStates.map((e: any) => SBookState.fromJSON(e))
        : [],
    };
  },

  toJSON(message: SBook): unknown {
    const obj: any = {};
    if (message.Author !== "") {
      obj.Author = message.Author;
    }
    if (message.Description !== "") {
      obj.Description = message.Description;
    }
    if (message.BookStates?.length) {
      obj.BookStates = message.BookStates.map((e) => SBookState.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SBook>, I>>(base?: I): SBook {
    return SBook.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SBook>, I>>(object: I): SBook {
    const message = createBaseSBook();
    message.Author = object.Author ?? "";
    message.Description = object.Description ?? "";
    message.BookStates = object.BookStates?.map((e) => SBookState.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const SBookMove: MessageFns<SBookMove> = {
  encode(message: SBookMove, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.Move !== 0) {
      writer.uint32(8).int32(message.Move);
    }
    if (message.Evaluation !== 0) {
      writer.uint32(16).int32(message.Evaluation);
    }
    if (message.Weight !== 0) {
      writer.uint32(24).int32(message.Weight);
    }
    if (message.NextStateId !== 0) {
      writer.uint32(32).int32(message.NextStateId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SBookMove {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSBookMove();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.Move = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.Evaluation = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.Weight = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.NextStateId = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SBookMove {
    return {
      Move: isSet(object.Move) ? globalThis.Number(object.Move) : 0,
      Evaluation: isSet(object.Evaluation) ? sBookMoveEvaluationFromJSON(object.Evaluation) : 0,
      Weight: isSet(object.Weight) ? globalThis.Number(object.Weight) : 0,
      NextStateId: isSet(object.NextStateId) ? globalThis.Number(object.NextStateId) : 0,
    };
  },

  toJSON(message: SBookMove): unknown {
    const obj: any = {};
    if (message.Move !== 0) {
      obj.Move = Math.round(message.Move);
    }
    if (message.Evaluation !== 0) {
      obj.Evaluation = sBookMoveEvaluationToJSON(message.Evaluation);
    }
    if (message.Weight !== 0) {
      obj.Weight = Math.round(message.Weight);
    }
    if (message.NextStateId !== 0) {
      obj.NextStateId = Math.round(message.NextStateId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SBookMove>, I>>(base?: I): SBookMove {
    return SBookMove.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SBookMove>, I>>(object: I): SBookMove {
    const message = createBaseSBookMove();
    message.Move = object.Move ?? 0;
    message.Evaluation = object.Evaluation ?? 0;
    message.Weight = object.Weight ?? 0;
    message.NextStateId = object.NextStateId ?? 0;
    return message;
  },
};

function createBaseSBookEval(): SBookEval {
  return { EvaluationValue: 0, Depth: 0, SelDepth: 0, Nodes: 0n, Variation: "", EngineName: "" };
}

export const SBookEval: MessageFns<SBookEval> = {
  encode(message: SBookEval, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.EvaluationValue !== 0) {
      writer.uint32(8).int32(message.EvaluationValue);
    }
    if (message.Depth !== 0) {
      writer.uint32(16).int32(message.Depth);
    }
    if (message.SelDepth !== 0) {
      writer.uint32(24).int32(message.SelDepth);
    }
    if (message.Nodes !== 0n) {
      if (BigInt.asIntN(64, message.Nodes) !== message.Nodes) {
        throw new globalThis.Error(
          "value provided for field message.Nodes of type int64 too large",
        );
      }
      writer.uint32(32).int64(message.Nodes);
    }
    if (message.Variation !== "") {
      writer.uint32(42).string(message.Variation);
    }
    if (message.EngineName !== undefined && message.EngineName !== "") {
      writer.uint32(50).string(message.EngineName);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SBookEval {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSBookEval();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.EvaluationValue = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.Depth = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.SelDepth = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.Nodes = reader.int64() as bigint;
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.Variation = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.EngineName = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SBookEval {
    return {
      EvaluationValue: isSet(object.EvaluationValue)
        ? globalThis.Number(object.EvaluationValue)
        : 0,
      Depth: isSet(object.Depth) ? globalThis.Number(object.Depth) : 0,
      SelDepth: isSet(object.SelDepth) ? globalThis.Number(object.SelDepth) : 0,
      Nodes: isSet(object.Nodes) ? BigInt(object.Nodes) : 0n,
      Variation: isSet(object.Variation) ? globalThis.String(object.Variation) : "",
      EngineName: isSet(object.EngineName) ? globalThis.String(object.EngineName) : "",
    };
  },

  toJSON(message: SBookEval): unknown {
    const obj: any = {};
    if (message.EvaluationValue !== 0) {
      obj.EvaluationValue = Math.round(message.EvaluationValue);
    }
    if (message.Depth !== 0) {
      obj.Depth = Math.round(message.Depth);
    }
    if (message.SelDepth !== 0) {
      obj.SelDepth = Math.round(message.SelDepth);
    }
    if (message.Nodes !== 0n) {
      obj.Nodes = message.Nodes.toString();
    }
    if (message.Variation !== "") {
      obj.Variation = message.Variation;
    }
    if (message.EngineName !== undefined && message.EngineName !== "") {
      obj.EngineName = message.EngineName;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SBookEval>, I>>(base?: I): SBookEval {
    return SBookEval.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SBookEval>, I>>(object: I): SBookEval {
    const message = createBaseSBookEval();
    message.EvaluationValue = object.EvaluationValue ?? 0;
    message.Depth = object.Depth ?? 0;
    message.SelDepth = object.SelDepth ?? 0;
    message.Nodes = object.Nodes ?? 0n;
    message.Variation = object.Variation ?? "";
    message.EngineName = object.EngineName ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}